ice_rest/requirements.txt
cvxopt
imbalanced-learn
nltk
numpy
pandas
xlrd
pydash
scikit-learn
scipy
sklearn-pandas
pymongo
locustio
gunicorn
futures
click
falcon_cors
redis
celery	
unidecode
msgpack-python
dill
eventlet
pystruct
pyro4
spacy
gitpython
minio
pycrypto
sklearn_crfsuite
stanfordcorenlp
python-keycloak
property
cachetools
quantulum3
moment
word2number
datefinder
requests
